renamed fdpsignal.py to signal.py
