+++ fdp/classes/signal.py
