+++ fdp/classes/utilities.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
 
-from . import container, fdpsignal, shot
+from . import container, signal, shot
 
 def isContainer(obj):
-    return issubclass(obj.__class__, container.Container) and 'Container' in repr(type(obj))
+    return issubclass(type(obj), container.Container) and 'Container' in str(type(obj))
 
 def isSignal(obj):
-    return issubclass(obj.__class__, fdpsignal.Signal) and 'Signal' in repr(type(obj))
+    return issubclass(type(obj), signal.Signal) and 'Signal' in str(type(obj))
 
 def isAxis(obj):
-    return issubclass(obj.__class__, fdpsignal.Signal) and 'Axis' in repr(type(obj))
+    return issubclass(type(obj), signal.Signal) and 'Axis' in str(type(obj))
 
 def isShot(obj):
-    return issubclass(obj.__class__, shot.Shot) and 'Shot' in repr(type(obj))
+    return issubclass(type(obj), shot.Shot) and 'Shot' in str(type(obj))
